requirements.txt
numpy
scipy
chardet
numexpr
tables
pandas-0.13.1
